% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_con.R
\name{db_con}
\alias{db_con}
\title{Connect to a database}
\usage{
db_con(dbname = "residents_per_sector")
}
\arguments{
\item{dbname}{The database name}
}
\value{
a data frame
}
\description{
Connect to a database
}
\examples{

db_con("residents_per_sector")

}
