% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_resident_data}
\alias{get_resident_data}
\title{Return resident data}
\usage{
get_resident_data(data_table = FALSE)
}
\arguments{
\item{data_table}{Logical. Return a \code{data.table}?}
}
\value{
Either a \code{tibble} or a \code{data.table}
}
\description{
Return resident data
}
\examples{

get_resident_data(data_table = TRUE)

}
