% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_avalanche}
\alias{theme_avalanche}
\title{AVALANCHE ggplot2 theme}
\usage{
theme_avalanche(base_size = 14, ...)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:ggtheme]{ggplot2::theme_minimal()}}}
}
\value{
a ggplot theme
}
\description{
Minimalistic ggplot themes for use on AVALANCHE reports.
}
\examples{

ggplot2::qplot(iris$Sepal.Length) +
  theme_avalanche()
}
